requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.1
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-stubs==0.3.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
python-decouple==3.1
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
